@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Providers;
 
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Route;
+use Illuminate\Routing\Router;
 use function base_path;
 
 final class RouteServiceProvider extends ServiceProvider
@@ -18,12 +18,13 @@ final class RouteServiceProvider extends ServiceProvider
 
     public function map(): void
     {
-        Route::prefix('api')
+        $router = app()->get(Router::class);
+        $router->prefix('api')
             ->middleware('api')
             ->namespace($this->namespace)
             ->group(base_path('routes/api.php'));
 
-        Route::middleware('web')
+        $router->middleware('web')
             ->namespace($this->namespace)
             ->group(base_path('routes/web.php'));
     }
